Fix range-start and range-end overflow.
The rectangular range indicators were moved aside to allow the rounded
range-edge indicators to be visible. However, this caused the rectangular indicators
to overflow outside the calendar boundaries, as well as cover adjecent date-items.

Moving aside the indicators 50% means it will overflow 50% of its width. To solve
the issue we make the indicator's width  50% of its original width. Now there is
nothing left to overflow.